requirements.txt
python-dotenv
xhtml2pdf
plotly_express
kaleido
boto3